Major code & warning cleanup.
Removed all warnings except a deprecation one and added two others for things that bothered me :
- converting an image for use in an OpenGL context (which I have no idea how to fix)
- not observing screen resolution changes
- hardcoding the "No Sound" localizable string from Quicksilver

+++ Abracadabra-App/ACGestureDisplayView.h
@@ -10,21 +10,16 @@
 #import <Cocoa/Cocoa.h>
 #import "ACGesture.h"
 
-NSPoint ACUnitPointInFrame(NSPoint p,NSRect r);
-NSPoint ACCenteredUnitPointInFrame(NSPoint p,NSRect r);
-NSPoint ACUnitPointWithCenterAndScale(NSPoint p,NSPoint c,float scale);
-void ACDrawDotAtPoint(NSPoint p,float r);
+NSPoint ACUnitPointInFrame(NSPoint p, NSRect r);
+NSPoint ACCenteredUnitPointInFrame(NSPoint p, NSRect r);
+NSPoint ACUnitPointWithCenterAndScale(NSPoint p, NSPoint c, CGFloat scale);
+void ACDrawDotAtPoint(NSPoint p, CGFloat r);
 	
-@interface ACGestureDisplayView : NSView
-{
+@interface ACGestureDisplayView : NSView {
 	ACGesture *gesture;
 	NSArray *events;
 	BOOL drawToScale;
 }
-
 - (ACGesture *)gesture;
 - (void)setGesture:(ACGesture *)newGesture;
-
-
-
 @end
